+++ slo/src/TopicService/Program.cs
@@ -1,0 +1,4 @@
+﻿using Internal;
+using TopicService;
+
+return await Cli.Run(new SloTopicContext(), args);