+++ src/Ydb.Sdk/src/Services/Topic/Reader/InternalBatchMessages.cs
@@ -15,7 +15,10 @@ internal class InternalBatchMessages<TValue>
     private int _startMessageDataIndex;
 
     private int OriginalMessageCount => _batch.MessageData.Count;
-    private bool IsActive => _startMessageDataIndex < OriginalMessageCount && _readerSession.IsActive;
+
+    private bool IsActive => _partitionSession.IsActive &&
+                             _readerSession.IsActive &&
+                             _startMessageDataIndex < OriginalMessageCount;
 
     public InternalBatchMessages(
         StreamReadMessage.Types.ReadResponse.Types.Batch batch,
@@ -97,7 +100,8 @@ internal class InternalBatchMessages<TValue>
             batch: messages,
             readerSession: _readerSession,
             offsetsRange: offsetsRangeBatch,
-            partitionSessionId: _partitionSession.PartitionSessionId
+            partitionSessionId: _partitionSession.PartitionSessionId,
+            producerId: _batch.ProducerId
         );
 
         return true;